Количество результатов теперь int

+++ GeoCoding.GeoCodingService/Data/GeoCod.cs
@@ -28,6 +28,6 @@
         /// <summary>
         /// Количество результатов по искомому адресу
         /// </summary>
-        public byte CountResult { get; set; }
+        public int CountResult { get; set; }
     }
 }
